+++ sources/GradientButton.m
@@ -30,14 +30,14 @@
 	
 	[NSGraphicsContext saveGraphicsState];
 	
-	float startWhite = enable ? (highlighted ? .65 : .99) : .93;
-	float endWhite = enable ? (highlighted ? .62 : .95) : .95;
-	float baseWhite = enable ? (highlighted ? .59 : .9) : .92;
+	CGFloat startWhite = enable ? (highlighted ? .65 : .99) : .93;
+	CGFloat endWhite = enable ? (highlighted ? .62 : .95) : .95;
+	CGFloat baseWhite = enable ? (highlighted ? .59 : .9) : .92;
 	
 	NSRect bounds = frame;
 	bounds = NSMakeRect(bounds.origin.x+1, bounds.origin.y, bounds.size.width-2, bounds.size.height-1);
 	
-	float halfHeight = ceil(bounds.size.height/2);
+	CGFloat halfHeight = ceil(bounds.size.height/2);
 	NSRect highlight = NSMakeRect(bounds.origin.x, bounds.origin.y, bounds.size.width, halfHeight);
 	NSColor* startColor = [NSColor colorWithDeviceWhite:startWhite alpha:1.0];
 	NSColor* endColor = [NSColor colorWithDeviceWhite:endWhite alpha:1.0];
